"warning fixes" commit introduced warnings; fix them. (And I refuse to use PRI macros.)

+++ GCN_Memcard_Manager/Src/MemoryCards/GCMemcard.cpp
@@ -51,12 +51,12 @@ GCMemcard::GCMemcard(const char *filename, bool forceCreation, bool sjis, u16 _s
 		auto size = mcdFile.GetSize() - mci_offset;
 		if (size < MC_FST_BLOCKS*BLOCK_SIZE)
 		{
-			PanicAlertT("%s failed to load as a memorycard \nfile is not large enough to be a valid memory card file (0x%x bytes)", filename, size);
+			PanicAlertT("%s failed to load as a memorycard \nfile is not large enough to be a valid memory card file (0x%x bytes)", filename, (unsigned) size);
 			return;
 		}
 		if (size % BLOCK_SIZE)
 		{
-			PanicAlertT("%s failed to load as a memorycard \n Card file size is invalid (0x%x bytes)", filename, size);
+			PanicAlertT("%s failed to load as a memorycard \n Card file size is invalid (0x%x bytes)", filename, (unsigned) size);
 				return;
 		}
 
@@ -71,7 +71,7 @@ GCMemcard::GCMemcard(const char *filename, bool forceCreation, bool sjis, u16 _s
 			case MemCard2043Mb:
 				break;
 			default:
-				PanicAlertT("%s failed to load as a memorycard \n Card size is invalid (0x%x bytes)", filename, size);
+				PanicAlertT("%s failed to load as a memorycard \n Card size is invalid (0x%x bytes)", filename, (unsigned) size);
 				return;
 		}
 		if  (mci_offset)
